IGNITE-16705 / move InFlightFutures to ignite-core
