+++ modules/core/src/test/java/org/apache/ignite/internal/future/InFlightFuturesTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.compute;
+package org.apache.ignite.internal.future;
 
 import static java.util.Collections.singleton;
 import static java.util.stream.Collectors.toSet;
